chapter-2 | rename examples to chapter-2 for better segregation
