+++ chapter-2/person/person_test.go
@@ -4,7 +4,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/HiteshRepo/Modern-API-Design-with-gRPC/examples/person"
+	"github.com/HiteshRepo/Modern-API-Design-with-gRPC/chapter-2/person"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
